requirements.txt
cmake
onnx
torch==1.10
torchvision==0.11.1
numpy
pandas
Pillow
ipywidgets
opencv-python
tqdm
imantics
sklearn
shapely
PyQt5
'git+https://github.com/facebookresearch/fvcore'
'git+https://github.com/facebookresearch/detectron2.git'